Refactored computer lifecylce module and functions and some error handling addition
